% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{break.group.into.families}
\alias{break.group.into.families}
\title{Assign each family inside a group to separate new groups}
\usage{
break.group.into.families(group)
}
\arguments{
\item{group}{an integer: the group number of the group to be split}
}
\value{
the group numbers of the new groups
}
\description{
\code{break.group.into.families} allocates all genotypes in the
group that share the same parents to separate new groups.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
