% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{break.group.into.halfsib.families}
\alias{break.group.into.halfsib.families}
\title{Assign all genotypes inside a group that share one parent 
to separate new groups}
\usage{
break.group.into.halfsib.families(group, parent)
}
\arguments{
\item{group}{an integer: the group number of the group to be split}

\item{parent}{an integer: 1 to group by the first parent, 2 to group 
by the second}
}
\value{
the group numbers of the new groups
}
\description{
\code{break.group.into.halfsib.families} allocates all genotypes in the
group that share one same parent (either the same first or the same 
second parent) to separate new groups.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()}
}
\concept{grouping functions}
