% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{break.group.into.individuals}
\alias{break.group.into.individuals}
\title{Assign each genotype inside a group to a separate new group}
\usage{
break.group.into.individuals(group)
}
\arguments{
\item{group}{an integer: the group number of the group to be split}
}
\value{
the group numbers of the new groups
}
\description{
\code{break.group.into.individuals} allocates every genotype inside
a group to a separate group.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
