% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{break.group.with.probabilities}
\alias{break.group.with.probabilities}
\title{Randomly split all genotypes in a group between a set of buckets with provided 
probabilities.}
\usage{
break.group.with.probabilities(group, probabilities)
}
\arguments{
\item{group}{an integer: the group number of the group to be split}

\item{probabilities}{a vector of decimals: the probability a genotypes will be 
allocated to the first, second, third, etc. buckets.}
}
\value{
the group numbers of the new groups
}
\description{
Given a `probabilities` parameter of length n, n+1 groups will be created, with 
 1 - (sum of other probabilities) probability of being in the last group. If 
probabilities add up to 1, a warning will be raised but no extra group will 
be created.
That is, it is assumed that the sum of `probabilities` is less than 1, 
with the remaining probability falling towards a last group. The reason for this 
is that floating-point numbers have their imperfect accuracies, so we want to 
save the computer from having to guess if you meant a last probability with 
0.000001 probability or if you were trying to provide some decimals that added
up nicely to 1.
}
\details{
If the bucket capacities add up to more than 1, 
a warning will be raised, but the function will still run. Note, though, that
genotypes are allocated to buckets using cumulative probabilities, so groups 
with cumulative probabilities of more than 1 will get no members.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
