% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{change.allele.symbol}
\alias{change.allele.symbol}
\title{Change the visual representation of an allele}
\usage{
change.allele.symbol(from, to, marker = NA)
}
\arguments{
\item{from}{Character symbol that will be replaced. eg. "A"}

\item{to}{Character symbol that @a from will be replaced with. eg. "b"}

\item{marker}{Name of the genetic marker across which to perform this
replacement operation, or NA if the allele should be replaced
across all genetic markers.}
}
\description{
genomicSimulation represents alleles as single-character symbols. 
This function swaps occurrences of the allele represented
by one character with a different character, in a particular marker 
or across all markers.
}
\details{
All genotypes present in simulation will be affected.

If @a to is an allele that already exists, all distinction between 
alleles @a from and @a to will be irreversibly lost.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.local.GEBVs}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
