% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{change.eff.set.centres.of.allele}
\alias{change.eff.set.centres.of.allele}
\title{Add or replace centring values for one allele in a set of marker effects}
\usage{
change.eff.set.centres.of.allele(allele, to, eff.set = 1L, reset.centres = T)
}
\arguments{
\item{allele}{Character symbol of the allele whose effects across these markers
should be centred by these values.}

\item{to}{One of two things:
- A vector of numeric values. The length of the vector must be the same as the 
number of rows in the output of \code{see.marker.effects(eff.set)}, and the 
vector should be ordered so that the first entry is the centring value belonging
to the marker in the first row of \code{see.marker.effects(eff.set)}, the second
entry is the centring value belonging to the marker in the second row, and so forth.
Any/all previous centring values of this set of marker effects will be overwritten.
- or, a dataframe with two columns. The first column should contain marker names,
like the first column of \code{see.marker.effects(eff.set)}, and the second column
should contain the centring values corresponding to those markers. Centring values 
corresponding to markers names in the first column will be overwritten; centring
values of makrers not in the first column will be unchanged.}

\item{eff.set}{The ID of the effect set to be modified. Return value of load.data or load.effects}

\item{reset.centres}{If true, the centring values of the marker effect set after 
running this function will be exactly equal to the allele centring value times the allele effect. 
If false, the centring values after running this function will be equal
to their previous values plus the product of the new allele centring value and the allele effect.
This can be useful when you want to add centring values for multiple alleles:
run the function with reset.centres=T for the first allele, and reset.centres=F for 
the next alleles.}
}
\description{
The simulation can store a centring value for each marker in a set of 
marker effects. The centring value is a value that is subtracted from
its marker's estimated breeding value contribution when calculating breeding values.
That is, an estimated breeding value  is the sum (across all markers) of 
the marker effect of the first allele at that marker, plus 
the marker effect of the second allele at that marker,
minus that marker's centring value, if it has a centring value.
}
\details{
However, if your marker effects were calculated using GBLUP etc., it may be 
more familiar to think of subtracting something from the count of copies of 
a particular allele, before multiplying that modified/centred "count" to the 
marker effect of that allele. If your marker effects were calculated using 
such a method (a count matrix centred by allele frequency), then use this function
rather than \code{change.eff.set.centres} to set the centring values for 
breeding value calculation.

The allele centring values passed to this function will be multiplied to the 
corresponding marker effect of that allele at that marker, before being 
set as/added to (depending on the value of \code{reset.centres}) the 
centring value for the marker.

A marker effect set loaded from a file with \code{load.data} or \code{load.effects}
will initially have all centring values set to 0. 

This function modifies the marker effect set. Breeding values calculated with this 
set of marker effects before this function call may be different to those calculated
with it after this function call.
}
\seealso{
Other loader functions: 
\code{\link{change.eff.set.centres}()},
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.effects}()},
\code{\link{load.genotypes}()},
\code{\link{load.map}()}
}
\concept{loader functions}
