% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{change.eff.set.centres}
\alias{change.eff.set.centres}
\title{Add or replace the centring values of a set of marker effects}
\usage{
change.eff.set.centres(to, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{to}{One of two things:
- A vector of numeric values. The length of the vector must be the same as the 
number of rows in the output of \code{see.marker.effects(eff.set)}, and the 
vector should be ordered so that the first entry is the centring value belonging
to the marker in the first row of \code{see.marker.effects(eff.set)}, the second
entry is the centring value belonging to the marker in the second row, and so forth.
Any/all previous centring values of this set of marker effects will be overwritten.
- or, a dataframe with two columns. The first column should contain marker names,
like the first column of \code{see.marker.effects(eff.set)}, and the second column
should contain the centring values corresponding to those markers. Centring values 
corresponding to markers names in the first column will be overwritten; centring
values of makrers not in the first column will be unchanged.}

\item{eff.set}{The ID of the effect set to be modified. Return value of load.data or load.effects.
If not otherwise specified, uses the earliest-loaded marker effect set in this simulation.
For legacy reasons, this parameter can be referred to as either `effect.set` or `eff.set`.
If `effect.set` is set to a non-default value, it takes precedence over `eff.set`.}
}
\description{
The simulation can store a centring value for each marker in a set of 
marker effects. The centring value is a value that is subtracted from
its marker's estimated breeding value contribution when calculating breeding values.
That is, an estimated breeding value  is the sum (across all markers) of 
the marker effect of the first allele at that marker, plus 
the marker effect of the second allele at that marker,
minus that marker's centring value, if it has a centring value.
}
\details{
A marker effect set loaded from a file with \code{load.data} or \code{load.effects}
will initially have all centring values set to 0. 

This function modifies the marker effect set. Breeding values calculated with this 
set of marker effects before this function call may be different to those calculated
with it after this function call.
}
\seealso{
Other loader functions: 
\code{\link{change.eff.set.centres.of.allele}()},
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.effects}()},
\code{\link{load.genotypes}()},
\code{\link{load.map}()}
}
\concept{loader functions}
