% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{change.label.by.amount}
\alias{change.label.by.amount}
\title{Add/subtract from a custom label}
\usage{
change.label.by.amount(label, amount, group = NA)
}
\arguments{
\item{label}{the label number of the label to change.}

\item{amount}{the integer change to modify the label value for the chosen
genotypes. eg. 1 for incrementing label values by 1, -3 to subtract 3 from
all label values.}

\item{group}{NA or 0 to apply this increment to all genotypes in simulation memory,
or a group number or vector of group numbers to apply the incrememnt to only
members of those groups.}
}
\value{
0 on success
}
\description{
\code{change.label.by.amount} changes the values of a custom
label for all genotypes or for members of group(s) by a given increment amount, 
positive or negative.
}
\details{
Multiple group input supported.
}
\seealso{
Other label functions: 
\code{\link{change.label.to.this}()},
\code{\link{change.label.to.values}()},
\code{\link{change.names.to.values}()},
\code{\link{delete.label}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.label}()}
}
\concept{label functions}
