% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{change.label.to.this}
\alias{change.label.to.this}
\title{Set a flat value for a custom label}
\usage{
change.label.to.this(label, value, group = NA)
}
\arguments{
\item{label}{the label number of the label to set.}

\item{value}{an integer to which the label value of all genotypes selected 
by the `group` parameter will be changed.}

\item{group}{NA or 0 to change all genotypes in simulation memory,
or a group number or vector of group numbers to change only
members of those groups.}
}
\value{
0 on success
}
\description{
\code{change.label.to} changes the value of a custom label for all 
genotypes or for members of group(s) to a certain value.
}
\details{
Multiple group input supported.
}
\seealso{
Other label functions: 
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.values}()},
\code{\link{change.names.to.values}()},
\code{\link{delete.label}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.label}()}
}
\concept{label functions}
