% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{change.label.to.values}
\alias{change.label.to.values}
\title{Set values of a custom label}
\usage{
change.label.to.values(label, values, group = NA, startIndex = NA, skip = 0)
}
\arguments{
\item{label}{the label number of the label to change.}

\item{values}{A vector of integer values to copy into the label.}

\item{group}{NA or 0 to apply the values to all genotypes in simulation memory,
or a group number to apply the values to only members of that group.}

\item{startIndex}{A parameter equivalent to `skip`. Because of the potential of
the name to be confusing, it is recommended you use `skip` instead (the equivalence
is `skip = startIndex - 1`). `startIndex` does not refer to an internal
genomicSimulation genotype index/position, it refers to the index 
of the group member that should get the first value from the vector `values`
if the group members were stored in an R vector and used R's 1-based indexing.}

\item{skip}{Number of group members to skip before the member of the group who 
should get the first value from the vector `values`.}
}
\description{
Changes the values of a custom label for all 
genotypes or for members of group(s) to a sequence of values.
}
\details{
The function skips the first `skip` candidate genotypes in `group`, then
copies each entry in `values` into the appropriate custom label of the 
next sequential member of `group`.

If the number of values provided is larger than the number of candidates in 
the group minus `skip`, then trailing  values are ignored. 
If the number of values provided is smaller than the number of candidates in 
the group minus `skip`, then extra candidates' labels are left unchanged.
}
\seealso{
Other label functions: 
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.names.to.values}()},
\code{\link{create.new.label}()},
\code{\link{delete.label}()}
}
\concept{label functions}
