% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-loaders.R
\name{change.names.to.values}
\alias{change.names.to.values}
\title{Set the names of genotypes in simulation memory}
\usage{
change.names.to.values(values, group = NA, startIndex = 0)
}
\arguments{
\item{values}{A vector of strings to copy into the genotypes' name slots.}

\item{group}{NA or 0 to apply names to all genotypes in simulation memory,
or a group number to apply the names to only members of that group.}

\item{startIndex}{The position within the list of genotypes identified by `group`
and sorted by their index that the first entry of `values` should be copied into.}
}
\value{
0 on success
}
\description{
\code{change.names.to.values} changes the names of members of group(s) or
genotypes in general to a sequence of values.
}
\details{
The function orders the genotypes selected by the `group` parameter by their
index, finds the `startIndex`th genotype (counting up from 0), then pastes
the vector `values` in order into the names of 
the genotypes from there on.

If the vector of values provided is longer than the number of genotypes after
the `startIndex`th that are selected by the `group` parameter, then trailing 
values are ignored. 
If the vector of values provided is shorter than the number of genotypes after
the `startIndex`th that are selected by the `group` parameter, then trailing 
genotype names are left unchanged.
}
\seealso{
Other label functions: 
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.label.to.values}()},
\code{\link{delete.label}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.label}()}
}
\concept{label functions}
