% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{clear.simdata}
\alias{clear.simdata}
\title{Clear the internal storage of all data}
\usage{
clear.simdata()
}
\description{
\code{clear.simdata} frees and deletes all data stored in the
package's internal SimData struct (sim.data$p). Returns 0 on success.
}
\seealso{
Other loader functions: 
\code{\link{load.data}()},
\code{\link{load.more.effects}()},
\code{\link{load.more.genotypes}()}
}
\concept{loader functions}
