% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{combine.groups}
\alias{combine.groups}
\title{Assign multiple groups' worth of genotypes to a single group}
\usage{
combine.groups(groups)
}
\arguments{
\item{groups}{A vector containing the group numbers of groups 
that are to be combined}
}
\value{
the group number of the combined group
}
\description{
\code{combine.group} combines the members of multple groups
into a single group. This is useful for allowing certain crossing
functions to act on these genotypes together.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
