% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{create.markerblocks.from.chrsplit}
\alias{create.markerblocks.from.chrsplit}
\title{Create marker blocks by splitting chromosomes by length}
\usage{
create.markerblocks.from.chrsplit(n.blocks.per.chr, map = 0L)
}
\arguments{
\item{n.blocks.per.chr}{Number of blocks into which each chromosome should be split.
If this is 1, then each block corresponds to one chromosome. If it is more than 1, 
then each chromosome is divided into this many equal-length segments, and a block
created containing the markers that appear in each of those segments.}

\item{map}{identifier of a genetic map that defines the chromosomes/linkage groups 
that are to be split by length to make the blocks. By default, it will use the 
oldest map that exists in the simulation.}
}
\value{
a genomicSimulation MarkerBlocks handle that can be passed to 
\code{see.local.GEBVs} or \code{save.local.GEBVs}
}
\description{
This function is for defining what markers are grouped together when calculating 
local GEBVs with \link{see.local.GEBVs} or \link{save.local.GEBVs}. Whereas 
\link{create.markerblocks} gives you freedom to group markers however you want,
this function provides a simpler interface to creating blocks where blocks 
are created based on lengths of distance in a chromosome.
}
\details{
It divides each chromosome in the map into a certain number of equal-length
segments (equal length according to centimorgan distances), and from each 
segment creates a block representing all the markers that appear in that chromosome region.
It is therefore possible that some blocks will have no markers in them, if there
are 'gaps' in the distribution of markers across the chromosome. The total number
of blocks created will be `number of chromosomes times n.blocks.per.chr`
}
