% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{create.markerblocks}
\alias{create.markerblocks}
\title{Create a set of marker blocks from a table of allocations}
\usage{
create.markerblocks(marker.allocations)
}
\arguments{
\item{marker.allocations}{a dataframe or list with two columns, the first 
containing marker names, the second containing a number representing the block
to which the marker should belong. The table does not need to contain every 
marker from the genetic map in the first column, and marker names can be repeated.
Unknown marker names will be ignored.
Block numbers in the second column can be 
any integers (they do not need to be sequential), as long as it is understood that 
markers with the same block number will be grouped together.}
}
\value{
a genomicSimulation MarkerBlocks handle that can be passed to 
\code{see.local.GEBVs} or \code{save.local.GEBVs}
}
\description{
This function is for defining what markers are grouped together when calculating 
local GEBVs with \link{see.local.GEBVs} or \link{save.local.GEBVs}. The user 
defines the allocations of individual markers to blocks. The list of markers 
in the genetic map which can be allocated can be accessed via \link{see.genetic.map}.
}
\seealso{
\link{create.markerblocks.from.chrsplit}
}
