% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{create.new.label}
\alias{create.new.label}
\title{Create a custom label}
\usage{
create.new.label(default)
}
\arguments{
\item{default}{the default value of the new custom label. All current genotypes
and all genotypes generated in future will have this value for this label.}
}
\value{
The label number of the new label created.
}
\description{
\code{make.label} creates a new custom label with the given default value.
}
\seealso{
Other label functions: 
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.label.to.values}()},
\code{\link{change.names.to.values}()},
\code{\link{delete.label}()}
}
\concept{label functions}
