% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{cross.all.pairs}
\alias{cross.all.pairs}
\title{OLD NAME | Performs crosses between every line and every other line in a group
in one direction}
\usage{
cross.all.pairs(
  group,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = NULL,
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\description{
! This is the old name for \code{make.all.unidirectional.crosses}. From genomicSimulation v0.2.5,
\code{make.all.unidirectional.crosses} is the recommended name over \code{cross.all.pairs}. 
\code{cross.all.pairs} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{make.all.unidirectional.crosses}
}
\keyword{internal}
