% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{cross.combinations}
\alias{cross.combinations}
\title{OLD NAME | Performs defined crosses as passed in as R vectors.}
\usage{
cross.combinations(
  first.parents,
  second.parents,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = "",
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\description{
! This is the old name for \code{make.targeted.crosses}. From genomicSimulation v0.2.5,
\code{make.targeted.crosses} is the recommended name over \code{cross.combinations}. 
\code{cross.combinations} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{make.targeted.crosses}
}
\keyword{internal}
