% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{cross.dc.combinations.file}
\alias{cross.dc.combinations.file}
\title{OLD NAME | Performs defined crosses between children of known parents as 
laid out in a file.}
\usage{
cross.dc.combinations.file(
  cross.file,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = "",
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\description{
! This is the old name for \code{make.double.crosses.from.file}. From genomicSimulation v0.2.5,
\code{make.double.crosses.from.file} is the recommended name over \code{cross.dc.combinations.file}. 
\code{cross.dc.combinations.file} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{make.double.crosses.from.file}
}
\keyword{internal}
