% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{cross.randomly.between}
\alias{cross.randomly.between}
\title{OLD NAME | Performs random crosses between two groups.}
\usage{
cross.randomly.between(
  group1,
  group2,
  cap1 = 0,
  cap2 = 0,
  n.crosses = 5,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = NULL,
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\description{
! This is the old name for \code{make.random.crosses.between}. From genomicSimulation v0.2.5,
\code{make.random.crosses.between} is the recommended name over \code{cross.randomly.between}. 
\code{cross.randomly.between} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{make.random.crosses.between}
}
\keyword{internal}
