% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{cross.randomly}
\alias{cross.randomly}
\title{OLD NAME | Performs random crosses between members of a group.}
\usage{
cross.randomly(
  group,
  n.crosses = 5,
  cap = 0,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = "",
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\description{
! This is the old name for \code{make.random.crosses}. From genomicSimulation v0.2.5,
\code{make.random.crosses} is the recommended name over \code{cross.randomly}. 
\code{cross.randomly} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{make.random.crosses}
}
\keyword{internal}
