% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-crossers.R
\name{cross}
\alias{cross}
\title{Perform a cross between two specific lines.}
\usage{
cross(
  parent1.index,
  parent2.index,
  offpsring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = NULL,
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE
)
}
\arguments{
\item{offpsring}{The number of offspring to produce from this pair of parents}

\item{retain}{A boolean, repesenting whether to save the generated
genotypes to memory or discard them. You may wish to discard them
but save to file if you are generating too many crosses to save into
memory.}

\item{give.names}{A boolean representing whether or not to produce names
for the new genotypes generated. The names produced would have format [name.prefix][id]}

\item{name.prefix}{A string. If give.names is TRUE, the id is concatenated to 
this to produce the name of each new genotype.}

\item{track.pedigree}{A boolean representing whether or not to save the ids
of the parents of each new genotype to the new genotype's pedigree. If this is 
false, the new genotype's pedigree is unknown.}

\item{give.ids}{A boolean representing whether or not to allocate each new 
genotype an id. If this is FALSE, the new genotype is 'invisible' to pedigree trackers and
even if the pedigree of its offspring is supposedly tracked, the pedigree trackers
will not be able to identify the progenitors of its offspring. Furthermore, if it is
false and names are generated using give.names, all names generated in the same group will
be the same. Probably you'd only have this FALSE if you were discarding the results or worried
about id overflow.}

\item{file.prefix}{A string representing the prefix of files produced if save.pedigree=TRUE, 
save.gebv=TRUE, or save.genotype=TRUE.}

\item{save.pedigree}{A boolean. If TRUE, saves the pedigree in recursive format of each
generated genotype to the file with filename "[file.prefix]-pedigree". This is a text file. 
Generated genotypes are saved progressively (up to 1000 at a time), so if the 
full result of a crosser function call will not fit in memory, this setting can allow
you to still get results.}

\item{save.gebv}{A boolean. If TRUE, saves the GEBVs of each generated genotype to the
file with filename "[file.prefix]-eff". This is a tab-separated text file. 
Valuse for generated genotypes are calculated and saved progressively (up to 1000 at a 
time), so if the full result of a crosser function call will not fit in memory, this 
setting can allow you to still get results.}

\item{save.genotype}{A boolean. If TRUE, saves the SNP/line matrix in regular format
(generated genotypes as rows, SNPs as columns) to the file with filename 
"[file.prefix]-genome". This is a tab-separated text file. 
Generated genotypes are saved progressively (up to 1000 at a time), so if the 
full result of a crosser function call will not fit in memory, this setting can allow
you to still get results.}
}
\value{
The group number of the group that the generated offspring were loaded into.
}
\description{
\code{cross} allows you to perform crosses between specific lines. It 
can be called in a for loop, looping over the rows of a dataframe, to 
achieve similar functionality to \code{\link{cross.combinations}} without
needing to save the combinations to be performed to a file.
}
\details{
The offpsring parameter represents the number of offspring to make from
this pair of parents.
}
\seealso{
Other crossing functions: 
\code{\link{cross.all.pairs}()},
\code{\link{cross.combinations}()},
\code{\link{cross.dc.combinations}()},
\code{\link{cross.randomly}()},
\code{\link{make.doubled.haploids}()},
\code{\link{self.n.times}()}
}
\concept{crossing functions}
