% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{define.matrix.format.details}
\alias{define.matrix.format.details}
\title{Manually specify the format for a genotype matrix file that will be loaded}
\usage{
define.matrix.format.details(
  has.header = NULL,
  markers.as.rows = NULL,
  cell.style = NULL
)
}
\arguments{
\item{has.header}{TRUE if the first line of the target file is a header row, 
FALSE if the first line of the target file is not a header row.}

\item{markers.as.rows}{TRUE if the rows of the genotype matrix represent genetic markers,
FALSE if the columns of the gentoype matrix represent genetic markers.}

\item{cell.style}{a string that will be used to identify the encoding of alleles in 
the body of the genotype matrix. String parsing is case-insensitive.
\itemize{
\item If the first character is "P", the target file stores alleles as ordered pairs of characters
(eg. "AA", "AT", "TA", "TT", "xy", "++")
\item If the first character is "/", the target file stores alleles as ordered pairs of 
characters separated by a forward slash character (eg. "A/A", "A/T", "T/A", "T/T", "x/y", "+/+")
\item If the first character is "C", the target file stores alleles as alternate allele counts
(ie. "0", "1", "2")
\item If the first character is "I", the target file stores alleles as IUPAC-encoded unordered pairs
(eg. "A"(=AA),"T"(=TT),"Y"(=TC or CT))
}}
}
\value{
a list suitable to be passed to the \code{format} parameter of \code{load.genotypes}
or \code{load.data}.
}
\description{
For use in generating the \code{format} parameter of
\link{load.genotypes} or \link{load.data}, when the genotype file 
being loaded is a genotype matrix.
}
\details{
It is not required that you manually specify all parameters, or none.
Whatever details of the file format are not manually specified
(that are not passed to this function or that are defined as NA) 
will be automatically detected by genomicSimulation when a genotype file is loaded
with this file format specification.
}
