% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.effect.set}
\alias{delete.effect.set}
\title{OLD NAME | Delete marker effect set(s)}
\usage{
delete.effect.set(effect.sets)
}
\description{
! This is the old name for \code{delete.effects}. It has been changed for consistency
with \code{load.effects}
}
\seealso{
\link{delete.effects}
}
\keyword{internal}
