% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{delete.group}
\alias{delete.group}
\title{Delete a group's worth of genotypes}
\usage{
delete.group(group)
}
\arguments{
\item{group}{an integer representing the group number of the group to be deleted}
}
\value{
0 on success. An error is raised on failure.
}
\description{
\code{delete.group} finds the genotypes of the given group and removes
them from the SimData object's storage. A message is printed explaining
how many genotypes were deleted.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
