% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.label}
\alias{delete.label}
\title{Delete custom label(s)}
\usage{
delete.label(labels)
}
\arguments{
\item{labels}{an vector containing the label numbers of the labels to be deleted}
}
\value{
0 on success. An error is raised on failure.
}
\description{
\code{delete.label} removes the given custom label(s) from all genotypes
and deletes them from simulation memory.
}
\seealso{
Other label functions: 
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.label.to.values}()},
\code{\link{change.names.to.values}()},
\code{\link{create.new.label}()}

Other deletor functions: 
\code{\link{clear.simdata}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.recombination.map}()}
}
\concept{deletor functions}
\concept{label functions}
