% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.recombination.map}
\alias{delete.recombination.map}
\title{OLD NAME | Delete recombination map(s)}
\usage{
delete.recombination.map(maps)
}
\description{
! This is the old name for \code{delete.map}. It has been changed for consistency
with \code{load.map}
}
\seealso{
\link{delete.map}
}
\keyword{internal}
