% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{find.crossovers}
\alias{find.crossovers}
\title{Identify historical crossover events}
\usage{
find.crossovers(parentage.file, out.file, window.size = 1, certainty = TRUE)
}
\arguments{
\item{parentage.file}{A string containing a filename. The file should
contain tab-separated, the name of an line and the name of its parents.
Parents do not have to match the parents tracked by its pedigree, but instead
are the ones from which the recombination frequencies will be calculated.}

\item{out.file}{A string containing a filename. The results of the analysis 
will be saved here as tab-separated matrix with child lines as the columns,
SNPs as the rows, and either a 0 (for undefined ancestry) or the id of the 
parent from which its alleles were inherited as the value in each cell.}

\item{window.size}{An odd integer representing the size of the window centered at
each SNP in the matrix over which to check which parent the alleles come from.}

\item{certainty}{A boolean. If TRUE, only SNPs where we can tell which parent the
alleles came from are filled in. If FALSE, uncertain SNPs are filled with the id of
the parent most recently identified as having given an allele.}
}
\value{
0 on success. The result will be printed to the file of name out.file
}
\description{
\code{find.crossovers} identifies alleles in genotypes that must
have come from one parent or another, and so estimates the sections
of the genotype that must have come from each parent.
}
\details{
**Experimental function: may be changed or removed at any time, and result validity is not promised.**
}
\seealso{
\code{\link{find.plot.crossovers}}, which performs the same calculations
but also produces a plot.
}
