% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicSimulation.R
\docType{package}
\name{genomicSimulation}
\alias{genomicSimulation}
\title{genomicSimulation: Crossing and genomic selection simulation tool}
\description{
Functions to simulate breeding programs on SNP marker genomes, and 
functions to calculate or perform selection based on GEBVs from provided effect values.
}
