% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-loaders.R
\name{load.data}
\alias{load.data}
\title{Create a new SimData object from data loaded from files}
\usage{
load.data(allele.file, map.file, effect.file = NULL, group.name = "Founders")
}
\arguments{
\item{allele.file}{A string containing a filename. The file should
contain a matrix of markers and alleles}

\item{map.file}{A string containing a filename. The file should contain
a linkage map for the markers loaded from allele.file}

\item{effect.file}{(optional) A string containing a filename. The
file should contain effect values for calculating GEBVs of a trait}

\item{group.name}{Name of the group to which the genotypes in the allele.file
will be loaded. Set this parameter to NULL to not use a name.}
}
\value{
The group number of the genotypes loaded from allele.file. This is
always 1 in the current implementation.
}
\description{
\code{load.data} returns the group number of the group the starting
data set was loaded into. It also sets up the markers, linkage map,
and GEBV calculators.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.different.effects}()},
\code{\link{load.more.genotypes}()}
}
\concept{loader functions}
