% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.different.effects}
\alias{load.different.effects}
\title{OLD NAME | Add a new set of marker effect values}
\usage{
load.different.effects(effect.file)
}
\description{
! This is the old name for \code{load.more.effects}. From genomicSimulation v0.2.5,
\code{load.more.effects} is the recommended name over \code{load.different.effects}. 
\code{load.different.effects} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{load.more.effects}
}
\keyword{internal}
