% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.genotypes}
\alias{load.genotypes}
\title{Load genotypes to the existing SimData object from a file}
\usage{
load.genotypes(allele.file)
}
\value{
The group number of the genotypes loaded from allele.file.
}
\description{
\code{load.genotypes} returns the group number of the group
that the new genotypes were loaded into. Only the alleles at markers
already tracked by the SimData are saved.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.effects}()},
\code{\link{load.map}()}
}
\concept{loader functions}
