% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.map}
\alias{load.map}
\title{Load an additional recombination map from a file}
\usage{
load.map(map.file)
}
\arguments{
\item{map.file}{A string containing a filename. The file should contain
a linkage map for the markers loaded from allele.file}
}
\value{
The identifier of the recombination map loaded from map.file
}
\description{
\code{load.map} returns the identifier of the new recombination
map. Only the markers already tracked by the SimData are included
in this map.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.effects}()},
\code{\link{load.genotypes}()}
}
\concept{loader functions}
