% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.more.genotypes}
\alias{load.more.genotypes}
\title{OLD NAME | Load more genotypes to the existing SimData object from a file}
\usage{
load.more.genotypes(allele.file)
}
\description{
! This is the old name for \code{load.genotypes}. From genomicSimulation v0.2.5,
\code{load.genotypes} is the recommended name over \code{load.more.genotypes}. 
\code{load.more.genotypes} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{load.genotypes}
}
\keyword{internal}
