% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{make.group.from.label.range}
\alias{make.group.from.label.range}
\title{OLD NAME | Assign genotypes with label values in a certain range to a new group}
\usage{
make.group.from.label.range(label, rangeLowEnd, rangeHighEnd, group = NA)
}
\description{
! This is the old name for \code{break.group.by.label.range}. From genomicSimulation v0.2.5,
\code{break.group.by.label.range} is the recommended name over \code{make.group.from.label.range}. 
\code{make.group.from.label.range} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{break.group.by.label.range}
}
\keyword{internal}
