% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{make.group.from.label}
\alias{make.group.from.label}
\title{OLD NAME | Assign genotypes with a certain label value to a new group}
\usage{
make.group.from.label(label, value, group = NA)
}
\description{
! This is the old name for \code{break.group.by.label.value}. From genomicSimulation v0.2.5,
\code{break.group.by.label.value} is the recommended name over \code{make.group.from.label}. 
\code{make.group.from.label} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{break.group.by.label.value}
}
\keyword{internal}
