% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{make.group}
\alias{make.group}
\title{Assign certain genotypes to a new group}
\usage{
make.group(indexes)
}
\arguments{
\item{indexes}{A vector of integers containing the indexes of the genotypes
to move to the new group. Use \code{get.group.data} to get these}
}
\value{
the group number of the new group
}
\description{
\code{make.group} allocates the genotypes with indexes in the vector
indexes to a single new group.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
