% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-loaders.R
\name{make.label}
\alias{make.label}
\title{Create a custom label}
\usage{
make.label(default)
}
\arguments{
\item{default}{the default value of the new custom label. All current genotypes
and all genotypes generated in future will have this value for this label.}
}
\value{
The label number of the new label created.
}
\description{
\code{make.label} creates a new custom label with the given default value.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
