% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.allele.counts}
\alias{save.allele.counts}
\title{Generate and save count matrices of alleles in the simulation to a file}
\usage{
save.allele.counts(filename, group = NULL, allele, markers.as.rows = TRUE)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NULL, will count and print all genotypes.
Otherwise, if a group of that number exists, save only lines that belong
to that group.}

\item{allele}{The first character of this string will be used as the 
allele to count.}

\item{markers.as.rows}{A logical representing the orientation of the matrix
in the output file. If TRUE, SNP markers will be rows of the matrix, and candidate 
lines will be columns. If FALSE, SNP markers will be columns of the matrix, and 
candidate lines will be rows.}
}
\description{
Counts the number of occurrences of a given
allele at each SNP for each candidate in the simulation or for all 
candidates in a group. Prints
the output to a file in a tab-separated matrix.
}
\details{
Candidate lines are represented in the output file by their names, if they
have names, or by their pedigree IDs if they are unnamed.
}
\seealso{
Other saving functions: 
\code{\link{change.allele.symbol}()},
\code{\link{save.GEBVs}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.chrsplit}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
