% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.GEBVs}
\alias{save.GEBVs}
\title{Save the GEBVs of genotypes currently saved to the SimData.}
\usage{
save.GEBVs(filename, group = NULL, eff.set = 1L)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NULL, will print all genotypes.
Otherwise, if a group of that number exists, save only lines that belong
to that group.}

\item{eff.set}{identifier for the set of marker effects with which to calculate
GEBVs.}
}
\value{
0 on success. On failure an error will be raised.
}
\description{
\code{save.GEBVs} calculates GEBVs for a set of genotypes using 
the current effect values saved to the SimData, then saves those GEBVs 
to a file.
}
\details{
The function prints its results to a tab-separated file, where each
row contains, in order, the genotype's ID, name, and calculated GEBV
}
\seealso{
Other saving functions: 
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.by.chr}()},
\code{\link{save.local.GEBVs.by.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
