% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.genotypes}
\alias{save.genotypes}
\title{Save genotypes/lines currently tracked by the simulation.}
\usage{
save.genotypes(filename, group = NULL, type = "R")
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NULL, will print all genotypes.
Otherwise, if a group of that number exists, save only lines that belong
to that group. Non-integers and negatives raise an error. Nonexistent 
groups result in empty files.}

\item{type}{The printing format. Use a string starting with 'R' or 'r' 
to save in regular format (SNPs as columns, lines as rows). Use a string
starting with 'T' or 't' to save in transposed format (SNPs as rows, lines
as columns). Both formats produce tab-separated matrices. The regular
format may be slightly faster because reads from the SimData are more
contiguous.}
}
\value{
0 on success. On failure an error will be raised.
}
\description{
\code{save.genotypes} saves the alleles at each SNP of the
lines stored in the SimData to a file.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.local.GEBVs}()},
\code{\link{save.pedigrees}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.genotype}()}
}
\concept{saving functions}
