% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.local.GEBVs.blocks.from.chrsplit}
\alias{save.local.GEBVs.blocks.from.chrsplit}
\alias{save.local.gebvs.blocks.from.chrsplit}
\title{Save the local GEBVs of each block in each selected line's haplotypes to a file,
using blocks created by slicing chromosomes into segments.}
\usage{
save.local.GEBVs.blocks.from.chrsplit(
  filename,
  n.blocks.per.chr,
  group = NULL,
  map = 0L,
  eff.set = 1L
)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{n.blocks.per.chr}{An integer containing the number of blocks each chromosome
will be divided into.}

\item{group}{Save only lines that belong to this group.}

\item{map}{identifier for the recombination map to use to calculate genetic 
distances with which to split the map. By default uses the oldest loaded map
currently active in simulation.}

\item{eff.set}{identifier for the set of marker effects with which to calculate 
local GEBVs. By default uses the first loaded set of effect values.}
}
\description{
\code{save.local.GEBVs.blocks.from.chrsplit} calculates GEBVs for each block of markers created by
splitting each chromosome into `n.blocks.per.chr` same-length blocks for the maternal
and paternal halves of its genotype. The results are saved to a file. The output file 
is formatted with the blocks as unlabelled columns and two rows for each genotype, named 
[line name]_1 and [line name]_2. The entries in this matrix are the calculated local 
GEBVs/block effects.
}
\seealso{
Other saving functions: 
\code{\link{change.allele.symbol}()},
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
