% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.local.GEBVs.by.chr}
\alias{save.local.GEBVs.by.chr}
\alias{save.local.gebvs.by.chr}
\title{OLD NAME | Save the local GEBVs of each block in each selected line's haplotypes to a file,
using blocks created by slicing chromosomes into segments.}
\usage{
save.local.GEBVs.by.chr(filename, n.blocks.per.chr, group = NULL, eff.set = 1L)
}
\description{
! This is the old name for \code{save.local.GEBVs.blocks.from.chrsplit}. From genomicSimulation v0.2.5,
\code{save.local.GEBVs.blocks.from.chrsplit} is the recommended name over \code{save.local.GEBVs.by.chr}. 
\code{save.local.GEBVs.by.chr} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{save.local.GEBVs.blocks.from.chrsplit}
}
\keyword{internal}
