% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.local.GEBVs.by.file}
\alias{save.local.GEBVs.by.file}
\alias{save.local.gebvs.by.file}
\title{OLD NAME | Save the local GEBVs of each block in each selected line's 
haplotypes to a file, using blocks taken from a file.}
\usage{
save.local.GEBVs.by.file(filename, block.file, group = NULL, eff.set = 1L)
}
\description{
! This is the old name for \code{save.local.GEBVs.blocks.from.file}. From genomicSimulation v0.2.5,
\code{save.local.GEBVs.blocks.from.file} is the recommended name over \code{save.local.GEBVs.by.file}. 
\code{save.local.GEBVs.by.file} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{save.local.GEBVs.blocks.from.file}
}
\keyword{internal}
