% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.local.GEBVs}
\alias{save.local.GEBVs}
\title{Calculate and save local breeding values of candidates in the simulation to a file}
\usage{
save.local.GEBVs(filename, blocks, group = NA, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NA or 0, will print all genotypes.
Otherwise, if a group of that number exists, save only local GEBVs of haplotypes
that belong to that group.}

\item{effect.set}{identifier for the set of marker effects to be used to calculate breeding values.
If not otherwise specified, uses the earliest-loaded marker effect set in this simulation.
For legacy reasons, you can also set this parameter using the name `eff.set`.}

\item{eff.set}{Legacy name for the parameter now called `effect.set`}
}
\description{
Calculates GEBVs for the maternal and paternal halves of each candidate's genotype
in each of a set of block of markers (for some set of marker blocks 
created by \code{create.markerblocks}). The output is saved as a tab-separated
file, where the marker blocks correspond to columns, and each pair of rows 
corresponds to a candidate's two haplotypes. Each entry in the matrix contains the
local breeding value/local GEBV of that candidate's haplotype for that block of 
markers.
}
\seealso{
Other saving functions: 
\code{\link{change.allele.symbol}()},
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
