% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{see.existing.groups}
\alias{see.existing.groups}
\title{Get a list of the groups currently existing in the SimData}
\usage{
see.existing.groups()
}
\value{
A vector containing the numbers of every group in the SimData that 
currently has members
}
\description{
\code{see.existing.groups} collects all groups that currently have genotypes
allocated to them and returns the group numbers as a vector.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
