% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.GEBVs}
\alias{see.GEBVs}
\title{Get breeding values of candidates in the simulation}
\usage{
see.GEBVs(group, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{group}{a vector of integers: the group number(s) of the group(s) to
return breeding values for}

\item{effect.set}{identifier for the set of marker effects to be used to calculate breeding values.
For legacy reasons, you can also set this parameter using the name `eff.set`.}

\item{eff.set}{Legacy name for the parameter now called `effect.set`}
}
\value{
A vector containing the breeding values for each member of each group.
}
\description{
This is an alternate name for \code{see.group.data(..., data.type="B")}. Therefore, 
it can return the breeding values of one or more groups, but does not work with 
\code{group=0L} to get the breeding values of all candidates in the simulation.
You can get the breeding values of all candidates in the simulation by running
\code{see.GEBVs(see.existing.groups()$Group)}.
}
\seealso{
Other data access functions: 
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
