% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.genetic.map}
\alias{see.genetic.map}
\title{View one of the genetic maps being used by the simulation}
\usage{
see.genetic.map(mapID = 0L)
}
\arguments{
\item{mapID}{the identifier of the specific genetic map to view, or zero to 
view the first-loaded/default genetic map.}
}
\value{
A dataframe containing three columns: the first, named "marker", 
containing marker names; the second, named "chr", containing chromosome numbers;
and the third, named "pos", containing positions in cM.
}
\description{
Access the list of markers of one of the genetic maps loaded in 
the simulation. The row order of the returned dataframe will 
match the order of the markers as stored by the simulation (eg.
will match the marker order in genotypes from see.group.data(data.type="G")).
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.marker.effects}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
