% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.group.gebvs}
\alias{see.group.gebvs}
\title{Get GEBVs of a group}
\usage{
see.group.gebvs(group)
}
\arguments{
\item{group}{an integer: the group number of the group to return the GEBVs of.}
}
\value{
a dataframe whose columns are the GEBVs of the group members, named "GEBV", and 
the indexes of the group members, named "i".
}
\description{
\code{see.group.gebvs} gives the R user a dataframe containing the GEBVs and a matching
index for each genotype in a group. This can be used in conjunction with \code{\link{make.group}}
to perform a custom selection on a group.
}
\section{Deprecation}{

In the history of this package, there was no way to get GEBVs using \code{\link{see.group.data}},
because that function only dealt with data that required no calculation step. However, it 
now includes that capability, and this function has become nothing but a wrapper. Deprecating it
and exposing the underlying call:
\code{data.frame("i"=see.group.data(...,data.type="X"),"GEBV"=see.group.data(...,data.type="BV")))}
demystifies this function and allows users to see how they can use \code{\link{see.group.data}} for 
flexible data access.
}

\section{Usual workflow}{

The suggested workflow involves using this function to get GEBVs and matching indexes,
performing some transformation on the GEBVs (eg environmental masking then
sorting), identifying some of the entries for selection, and passing
the indexes of those entries to \code{\link{make.group}} in order to split
them off into a new group.
}

\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{select.by.gebv}()}

Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.minimum.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()}
}
\concept{data access functions}
\concept{grouping functions}
