% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.group.gene.data}
\alias{see.group.gene.data}
\title{Get genotypes or allele counts of a group as a matrix.}
\usage{
see.group.gene.data(group, count.allele = NA_character_)
}
\arguments{
\item{group}{an integer: the group number of the group to have a look at}

\item{count.allele}{a character that will be used to identify the type of data
to be extracted.
If it is NA, then return the pair of alleles as a string eg "AT"
If it is a character, then return the count of that allele eg. 'A' -> 0 if 
no copies of A at that marker, 1 if heterozygous for A, 2 if "AA" at that marker.}
}
\description{
\code{see.group.gene.data} allows you to extract (markers)x(group members)
matrices of the genetics at each marker of some of the genotypes in simulation
For the moment only implemented in R, not C
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()}

Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
\concept{grouping functions}
