% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.local.GEBVs}
\alias{see.local.GEBVs}
\title{Get local GEBVs of candidates in the simulation}
\usage{
see.local.GEBVs(blocks, group, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{blocks}{external pointer for a marker blocks object created by \code{create.markerblocks}}

\item{group}{a vector of integers: the group number(s) of the group(s) to
return breeding values for}

\item{effect.set}{identifier for the set of marker effects to be used to calculate breeding values.
By default is 0, to use the oldest currently-loaded set of marker effects.
For legacy reasons, you can also set this parameter using the name `eff.set`.}
}
\value{
A matrix containing the local breeding values for each haplotype of 
each member of each group. Columns correspond to blocks, each pair of rows 
corresponds to a candidate's two haplotypes. Candidates are ordered the same 
as their ordering in the output of \code{see.group.data}.
}
\description{
Calculates GEBVs for the maternal and paternal halves of each candidate's genotype
in each of a set of block of markers (for some set of marker blocks 
created by \code{create.markerblocks}).
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
