% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.markerblocks}
\alias{see.markerblocks}
\title{View the allocations of markers to blocks in one MarkerBlocks object.}
\usage{
see.markerblocks(b)
}
\arguments{
\item{b}{An external pointer to a MarkerBlocks object (eg. the pointer returned
by create.markerblocks)}
}
\value{
A dataframe with columns "marker" and "block". The former contains marker 
names, the latter contains the corresponding block index.
}
\description{
View the allocations of markers to blocks in one MarkerBlocks object.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
