% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.minimal.GEBV}
\alias{see.minimal.GEBV}
\alias{see.minimal.gebv}
\title{Get the lowest-possible GEBV given the current loaded effect values.}
\usage{
see.minimal.GEBV(effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{effect.set}{identifier for the set of marker effects to use to calculate the
minimum GEBV. If not otherwise specified, uses the earliest-loaded marker effect set 
in this simulation.
For legacy reasons, you can also set this parameter using the name `eff.set`.}
}
\value{
The lowest possible GEBV
}
\description{
Allows you to extract the lowest GEBV score
possible according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
