% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.minimum.GEBV}
\alias{see.minimum.GEBV}
\alias{see.minimum.gebv}
\title{OLD NAME | Get the lowest-possible GEBV given the current loaded effect values.}
\usage{
see.minimum.GEBV(eff.set = 1L)
}
\description{
! This is the old name for \code{see.minimal.GEBV}. From genomicSimulation v0.2.5,
\code{see.minimal.GEBV} is the recommended name over \code{see.minimum.GEBV}. 
\code{see.minimum.GEBV} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{see.minimal.GEBV}
}
\keyword{internal}
