% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.optimal.GEBV}
\alias{see.optimal.GEBV}
\title{Get the ultimate/highest-possible GEBV given the current loaded values.}
\usage{
see.optimal.GEBV()
}
\value{
The highest possible GEBV
}
\description{
\code{see.optimal.GEBV} allows you to extract the optimal GEBV
according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.haplotype}()}
}
\concept{data access functions}
