% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.optimal.genotype}
\alias{see.optimal.genotype}
\title{Get a string containing the ultimate/highest-scoring genotype.}
\usage{
see.optimal.genotype()
}
\value{
A string containing the allele out of available alleles at that
marker that has the highest effect value. The string will be ordered in
genome order (lowest chromosome and lowest position to highest) according
to the map that was included on initialisation.
}
\description{
\code{see.optimal.genotype} allows you to extract the optimal genotype 
according to the current loaded effect values as an R string. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.GEBV}()}
}
\concept{data access functions}
