% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.optimal.haplotype}
\alias{see.optimal.haplotype}
\title{Get a string containing the ultimate/highest-scoring set of alleles.}
\usage{
see.optimal.haplotype()
}
\value{
A string containing the allele out of available alleles at that
marker that has the highest effect value. The string will be ordered in
genome order (lowest chromosome and lowest position to highest) according
to the map that was included on initialisation.
}
\description{
\code{see.optimal.haplotype} allows you to extract the optimal haplotype 
according to the current loaded effect values as an R string. An error is 
raised if no effect values have been loaded. The optimal genotype (assuming
only additive allele effects) is just the doubled version of this haplotype.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.minimum.GEBV}()},
\code{\link{see.optimal.GEBV}()}
}
\concept{data access functions}
