% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.optimal.possible.GEBV}
\alias{see.optimal.possible.GEBV}
\alias{see.optimal.possible.gebv}
\title{Get the ultimate/highest-possible GEBV given the pool of alleles available in
the current group.}
\usage{
see.optimal.possible.GEBV(group, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{group}{an integer: the group number of the group to have a look at.
Can be a vector of groups, in which case the calculation will be run independently
for each group.}

\item{effect.set}{identifier for the set of marker effects to use to calculate the
optimal possible GEBV. If not otherwise specified, uses the earliest-loaded 
marker effect set in this simulation.
For legacy reasons, you can also set this parameter using the name `eff.set`.}
}
\value{
The highest possible GEBV that can be created from alleles available
in the group.
}
\description{
Effectively, given the additive model of trait effects, this is 2 times the 
breeding value score of \code{see.optimal.possible.haplotype}.
}
\details{
An error is raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
