% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{select.by.gebv}
\alias{select.by.gebv}
\title{OLD NAME | Perform selection on a group by true GEBV.}
\usage{
select.by.gebv(
  from.group,
  low.score.best = FALSE,
  percentage = NULL,
  number = NULL,
  eff.set = 1L
)
}
\description{
! This is the old name for \code{break.group.by.GEBV}. From genomicSimulation v0.2.5,
\code{break.group.by.GEBV} is the recommended name over \code{select.by.gebv}. 
\code{select.by.gebv} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{break.group.by.GEBV}
}
\keyword{internal}
