% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{break.group.into.families}
\alias{break.group.into.families}
\title{Assign each family inside a group to separate new groups}
\usage{
break.group.into.families(group)
}
\arguments{
\item{group}{an integer: the group number of the group to be split}
}
\value{
the group numbers of the new groups
}
\description{
\code{break.group.into.families} allocates all genotypes in the
group that share the same parents to separate new groups.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()}
}
\concept{grouping functions}
