% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{change.label.to.values}
\alias{change.label.to.values}
\title{Set values of a custom label}
\usage{
change.label.to.values(label, values, group = NA, startIndex = 1)
}
\arguments{
\item{label}{the label number of the label to change.}

\item{values}{A vector of integer values to copy into the label.}

\item{group}{NA or 0 to apply the values to all genotypes in simulation memory,
or a group number to apply the values to only members of that group.}

\item{startIndex}{The position within the list of genotypes identified by `group`
and sorted by their index that the first entry of `values` should be copied into.
The first position in the list is considered index 1.}
}
\value{
0 on success
}
\description{
Changes the values of a custom label for all 
genotypes or for members of group(s) to a sequence of values.
}
\details{
The function orders the genotypes selected by the `group` parameter by their
index, finds the `startIndex`th genotype (counting up from 0), then copies
the vector `values` entry-by-entry into the appropriate label of 
the genotypes from there on.

If the vector of values provided is longer than the number of genotypes after
the `startIndex`th that are selected by the `group` parameter, then trailing 
values are ignored. 
If the vector of values provided is shorter than the number of genotypes after
the `startIndex`th that are selected by the `group` parameter, then trailing 
genotypes are left unchanged.
}
\seealso{
Other label functions: 
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.names.to.values}()},
\code{\link{create.new.label}()},
\code{\link{delete.label}()}
}
\concept{label functions}
