% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{clear.simdata}
\alias{clear.simdata}
\title{Clear the internal storage of all data}
\usage{
clear.simdata()
}
\description{
Frees and deletes all data stored in the
package's internal SimData struct (sim.data$p). Returns 0 on success.
}
\seealso{
Other loader functions: 
\code{\link{change.eff.set.centres.of.allele}()},
\code{\link{change.eff.set.centres}()},
\code{\link{load.data}()},
\code{\link{load.effects}()},
\code{\link{load.genotypes}()},
\code{\link{load.map}()}

Other deletor functions: 
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{delete.recombination.map}()}
}
\concept{deletor functions}
\concept{loader functions}
