% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.effect.set}
\alias{delete.effect.set}
\title{Delete marker effect set(s)}
\usage{
delete.effect.set(effect.sets)
}
\arguments{
\item{effect.sets}{an vector containing the effect set identifiers of the 
effect sets to be deleted}
}
\value{
0 on success. An error is raised on failure.
}
\description{
Removes the given marker effect set(s) 
from simulation memory.
}
\seealso{
Other deletor functions: 
\code{\link{clear.simdata}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()},
\code{\link{delete.recombination.map}()}
}
\concept{deletor functions}
