% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.group}
\alias{delete.group}
\title{Delete a group's worth of genotypes}
\usage{
delete.group(groups)
}
\arguments{
\item{groups}{an vector containing the group numbers of the groups to be deleted}
}
\value{
0 on success. An error is raised on failure.
}
\description{
Finds the genotypes of the given group and removes
them from the SimData object's storage. A message is printed explaining
how many genotypes were deleted.
}
\details{
Multiple group input supported.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()}

Other deletor functions: 
\code{\link{clear.simdata}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.label}()},
\code{\link{delete.recombination.map}()}
}
\concept{deletor functions}
\concept{grouping functions}
