% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.label}
\alias{delete.label}
\title{Delete custom label(s)}
\usage{
delete.label(labels)
}
\arguments{
\item{labels}{an vector containing the label numbers of the labels to be deleted}
}
\value{
0 on success. An error is raised on failure.
}
\description{
\code{delete.label} removes the given custom label(s) from all genotypes
and deletes them from simulation memory.
}
\seealso{
Other label functions: 
\code{\link{change.label.by.amount}()},
\code{\link{change.label.to.this}()},
\code{\link{change.label.to.values}()},
\code{\link{change.names.to.values}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.label}()}

Other grouping functions: 
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{make.group.from.label.range}()},
\code{\link{make.group.from.label}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{select.by.gebv}()}
}
\concept{grouping functions}
\concept{label functions}
