% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deletors.R
\name{delete.recombination.map}
\alias{delete.recombination.map}
\title{Delete recombination map(s)}
\usage{
delete.recombination.map(maps)
}
\arguments{
\item{maps}{an vector containing the map identifiers of the maps to be deleted}
}
\description{
Removes the given recombination map(s)
from simulation memory.
}
\seealso{
Other deletor functions: 
\code{\link{clear.simdata}()},
\code{\link{delete.effect.set}()},
\code{\link{delete.group}()},
\code{\link{delete.label}()}
}
\concept{deletor functions}
