% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicSimulation.R
\docType{package}
\name{genomicSimulation}
\alias{genomicSimulation}
\title{genomicSimulation: Stochastic simulation of breeding programs}
\description{
Functions to simulate breeding programs on (user-provided) sets of genetic markers.
}
