% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicSimulation.R
\docType{package}
\name{genomicSimulation}
\alias{genomicSimulation}
\title{genomicSimulation: Stochastic simulation of breeding programs}
\description{
Functions to simulate breeding programs on (user-provided) sets of genetic markers.
}
\details{
Latest package vignette: https://vllrs.github.io/genomicSimulation/doc/gSvignette.html

Older overview of package (some details may be out of date): https://www.biorxiv.org/content/10.1101/2021.12.12.472291v2

Package sources: https://github.com/vllrs/genomicSimulation
}
