% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.data}
\alias{load.data}
\title{Create a new SimData object from data loaded from files}
\usage{
load.data(allele.file, map.file, effect.file = NULL)
}
\arguments{
\item{allele.file}{A string containing a filename. The file should
contain a matrix of markers and alleles}

\item{map.file}{A string containing a filename. The file should contain
a linkage map for the markers loaded from allele.file}

\item{effect.file}{(optional) A string containing a filename. The
file should contain effect values for calculating GEBVs of a trait}
}
\value{
The group number of the genotypes loaded from allele.file. This is
always 1 in the current implementation. If an effect file was also loaded, the 
return value will be a list with two entries: groupNum, for the group number of the
first group loaded, and EffectID, for the identifier for the set of marker effects
loaded from the effect file.
}
\description{
\code{load.data} returns the group number of the group the starting
data set was loaded into. It also sets up the markers, linkage map,
and GEBV calculators.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.more.effects}()},
\code{\link{load.more.genotypes}()}
}
\concept{loader functions}
