% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.data}
\alias{load.data}
\title{Create a new SimData object from data loaded from files}
\usage{
load.data(genotype.file = NULL, map.file = NULL, effect.file = NULL)
}
\arguments{
\item{genotype.file}{A string containing a filename. The file should
contain a matrix of markers and alleles}

\item{map.file}{A string containing a filename. The file should contain
a linkage map for the markers loaded from allele.file}

\item{effect.file}{A string containing a filename. The
file should contain effect values for calculating GEBVs of a trait}
}
\value{
A list with entries $groupNum, $mapID, and $effectID, representing
the group number of the genotypes loaded from the genotype file, the 
recombination map identifier of the map loaded from the map.file, and 
the marker effect set identifier of the marker effects loaded from 
the effect.file respectively.
}
\description{
\code{load.data} sets up the simulation object with initial data
from the provided files. Either a genotype file or a map file, or 
both, should be provided, as a set of marker effects cannot be loaded
without a reference set of markers from either of the former two
files.
}
\details{
The list of markers tracked by the simulation is immutable after this
call. Therefore, the map file (or genotype file, if no map file 
is provided) should contain all markers you will wish to simulate. 

Additional founder genotypes, recombination maps, and marker effect
sets can be loaded after this call using @seealso \link{load.genotypes},
}
\seealso{
\link{load.map},

\link{load.effects}

See package vignette for file formats that the package can load.

Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.effects}()},
\code{\link{load.genotypes}()},
\code{\link{load.map}()}
}
\concept{loader functions}
