% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-loaders.R
\name{load.different.effects}
\alias{load.different.effects}
\title{Replace effect values}
\usage{
load.different.effects(effect.file)
}
\arguments{
\item{effect.file}{A string containing a filename. The file should
contain effect values for calculating GEBV for a trait}
}
\value{
0 on success.
}
\description{
\code{load.different.effects} returns 0. Only the alleles at SNPs
already tracked by the SimData are saved. The old effect values,
if they exist, will be replaced with these new ones.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.more.genotypes}()}
}
\concept{loader functions}
