% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{load.more.genotypes}
\alias{load.more.genotypes}
\title{Load more genotypes to the existing SimData object from a file}
\usage{
load.more.genotypes(allele.file)
}
\arguments{
\item{allele.file}{A string containing a filename. The file should
contain a matrix of markers and alleles}
}
\value{
The group number of the genotypes loaded from allele.file.
}
\description{
\code{load.more.genotypes} returns the group number of the group
that the new genotypes were loaded into. Only the alleles at SNPs
already tracked by the SimData are saved.
}
\seealso{
Other loader functions: 
\code{\link{clear.simdata}()},
\code{\link{load.data}()},
\code{\link{load.more.effects}()}
}
\concept{loader functions}
