% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-setup.R
\name{make.label}
\alias{make.label}
\title{OLD NAME | Create a custom label}
\usage{
make.label(default)
}
\description{
! This is the old name for \code{create.new.label}. From genomicSimulation v0.2.5,
\code{create.new.label} is the recommended name over \code{make.label}. 
\code{make.label} may become deprecated in future, when the package reaches 
stability.
}
\seealso{
\link{create.new.label}
}
\keyword{internal}
