% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-progression.R
\name{make.random.crosses.between}
\alias{make.random.crosses.between}
\title{Performs random crosses between two groups.}
\usage{
make.random.crosses.between(
  group1,
  group2,
  n.crosses = 5,
  cap1 = 0,
  cap2 = 0,
  map1 = 0L,
  map2 = 0L,
  offspring = 1,
  retain = TRUE,
  give.names = FALSE,
  name.prefix = NULL,
  track.pedigree = TRUE,
  give.ids = TRUE,
  file.prefix = "",
  save.pedigree = FALSE,
  save.gebv = FALSE,
  save.genotype = FALSE,
  map = 0L
)
}
\arguments{
\item{group1}{The group number from which to draw the first parent of the random 
crosses.}

\item{group2}{The group number from which to draw the second parent of the random 
crosses.}

\item{n.crosses}{The function will pick this many random pairs of parents
to cross.}

\item{cap1}{If nonzero, this is the maximum number of times each member of 
group1 can be used as the (first) parent of a cross.
Set to 0 for no restriction on the number of offspring per parent.}

\item{cap2}{If nonzero, this is the maximum number of times each member of 
group2 can be used as the (second) parent of a cross.
Set to 0 for no restriction on the number of offspring per parent.}

\item{map1}{The identifier for the recombination map with which gametes 
from members of \code{group1} will be generated. By default uses the oldest 
loaded map currently active in simulation.}

\item{map2}{The identifier for the recombination map with which gametes 
from members of \code{group2} will be generated. By default uses the oldest 
loaded map currently active in simulation}

\item{offspring}{The number of times to cross each randomly-chosen pair.}

\item{retain}{A logical value, repesenting whether to save the generated
genotypes to memory or discard them. You may wish to discard them
but save to file if you are generating too many crosses to save into
memory.}

\item{give.names}{A logical value representing whether or not to produce names
for the new genotypes generated. The names produced would have format [name.prefix][id]}

\item{name.prefix}{A string. If give.names is TRUE, the id is concatenated to 
this to produce the name of each new genotype.}

\item{track.pedigree}{A logical value representing whether or not to save the ids
of the parents of each new genotype to the new genotype's pedigree. If this is 
false, the new genotype's pedigree is unknown.}

\item{give.ids}{A logical value representing whether or not to allocate each new 
genotype an id. If this is FALSE, the new genotype is 'invisible' to pedigree trackers and
even if the pedigree of its offspring is supposedly tracked, the pedigree trackers
will not be able to identify the progenitors of its offspring. Furthermore, if it is
false and names are generated using give.names, all names generated in the same group will
be the same. Probably you'd only have this FALSE if you were discarding the results or worried
about id overflow.}

\item{file.prefix}{A string representing the prefix of files produced if save.pedigree=TRUE, 
save.gebv>0, or save.genotype=TRUE. If no file.prefix is provided, but at least one of 
those settings is TRUE, the default file prefix is "out".
save.pedigree, save.gebv, and save.genotype are known
as the "save-as-you-go" settings, which allow you to output data on the genotypes at the same 
time as they are being generated. The save-as-you-go settings can be used in conjunction with
retain=FALSE, if you want to generate results from a greater number of offspring than your 
system memory can hold.}

\item{save.pedigree}{A logical value. If TRUE, saves the pedigrees of each
generated genotype to the file with filename "[file.prefix]-pedigree.txt". 
Pedigrees are saved progressively (1000 at a time). They are saved in recursive 
predigree format - see function \code{save.pedigrees} or the package vignette for more details.}

\item{save.gebv}{An integer representing a marker effect set identifier, or 0.
If FALSE, 0, or negative, does not save GEBVs. If positive and corresponding to the identifier
of a loaded set of marker effects, saves the GEBVs of each generated genotype to the
file with filename "[file.prefix]-bv.txt". This is a tab-separated text file. 
Breeding values for generated genotypes are calculated and saved progressively (1000 at a 
time). See function \code{save.GEBVs} for more details on the output format. A reminder that 
if only one set of marker effects have been loaded, the identifier for that set of 
marker effects is 1.}

\item{save.genotype}{A logical value. If TRUE, saves the SNP matrix of the generated 
genotypes to the file with filename "[file.prefix]-genotype.txt". 
Generated genotypes are saved progressively (1000 at a time). The output file 
uses genetic markers as columns and genotypes as rows. See \code{save.genotypes}}

\item{map}{If \code{map1} and \code{map2} are not specified 
but \code{map} is, then the map specified by parameter \code{map} will be used for 
both map1 and map2.}
}
\value{
The group number of the new crosses produced, or 0 if they could not be
produced due to an invalid parent group number being provided.
}
\description{
Performs crosses where the first parent comes 
from one group and the second from another. It returns the group number of the group
that the new genotypes were loaded into.
}
\details{
For random crossing, the n.crosses parameter represents the number of 
random crosses to perform. Each cross will be repeated a number of times
determined by the offspring parameter. Each member of a group becomes 
a parent with equal probability.

If the user only wants one parent to be randomly chosen, the flag `set.parent1`
can be set to TRUE, in which case `group1` will instead be interpreted as the 
index of an individual genotype, to which members of the other group will be crossed.
(and similarly the second parent can be set using `set.parent2`. A reminder 
that parent order affects nothing but pedigree printing, so pick whichever.)

Parameters set.parent1 and set.parent2 are deprecated and removed!
Use make.group and combine.groups to temporarily move an individual to their own group
if you wish to cross randomly from a group to an individual.
}
\seealso{
Other crossing functions: 
\code{\link{make.all.unidirectional.crosses}()},
\code{\link{make.clones}()},
\code{\link{make.crosses.from.file}()},
\code{\link{make.double.crosses.from.file}()},
\code{\link{make.doubled.haploids}()},
\code{\link{make.random.crosses}()},
\code{\link{make.targeted.crosses}()},
\code{\link{self.n.times}()}
}
\concept{crossing functions}
