% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.allele.counts}
\alias{save.allele.counts}
\title{Generate and save count matrices for alleles in the SimData}
\usage{
save.allele.counts(filename, group = NULL, allele)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NULL, will count and print all genotypes.
Otherwise, if a group of that number exists, save only lines that belong
to that group.}

\item{allele}{The first character of this string will be used as the 
allele to count.}
}
\value{
0 on success. On failure an error will be raised.
}
\description{
\code{save.allele.counts} counts the number of occurences of a given
allele at each SNP of each genotype in the selected set, and prints
the output to a file.
}
\details{
The function prints its results to a tab-separated matrix file. The first line/
column header gives the name (or id if the genotype does not have a name allocated)
of each genotype that is having its counts printed. After that header row, all rows 
begin with the name of a SNP, followed by the number of the given allele at that
marker in the corresponding genotype from the header row.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.chrsplit}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
