% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.GEBVs}
\alias{save.GEBVs}
\title{Calculate and save breeding values of candidates in the simulation to a file}
\usage{
save.GEBVs(filename, group = NA, effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{group}{If not set/set to NA or 0, will print all genotypes.
Otherwise, if a group of that number exists, save only lines that belong
to that group.}

\item{effect.set}{identifier for the set of marker effects with which to calculate
GEBVs. If not otherwise specified, uses the earliest-loaded marker effect set in this simulation.
For legacy reasons, you can also set this parameter using the name `eff.set`.}

\item{eff.set}{Legacy name for the parameter now called `effect.set`}
}
\description{
The function prints its results to a tab-separated file with three columns 
(the genotype's ID, its name, and its calculated breeding value/GEBV).
}
\seealso{
Other saving functions: 
\code{\link{change.allele.symbol}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.blocks.from.file}()},
\code{\link{save.local.GEBVs}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
