% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.genome.model}
\alias{save.genome.model}
\title{Save the details of the genome that the SimData uses.}
\usage{
save.genome.model(filename)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}
}
\value{
0 on success. On failure an error will be raised.
}
\description{
\code{save.genome.model} saves the SNP names, their linkage map positions, 
and the effect values for calculating GEBVs (if applicable) to a file.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs}()},
\code{\link{save.pedigrees}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.genotype}()}
}
\concept{saving functions}
