% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-printers.R
\name{save.local.GEBVs.by.chr}
\alias{save.local.GEBVs.by.chr}
\title{Save the local GEBVs of each block in each selected line's haplotypes to a file,
using blocks created by slicing chromosomes into segments.}
\usage{
save.local.GEBVs.by.chr(filename, n.blocks.per.chr, group = NULL)
}
\arguments{
\item{filename}{A string containing a filename to which the output will
be written}

\item{n.blocks.per.chr}{An integer containing the number of blocks each chromosome
will be divided into.}

\item{group}{Save only lines that belong to this group.}
}
\value{
0 on success. On failure an error will be raised.
}
\description{
\code{save.local.GEBVs.by.chr} calculates GEBVs for each block of markers created by
splitting each chromosome into `n.blocks.per.chr` same-length blocks for the maternal
and paternal halves of its genotype. The results are saved to a file. The output file 
is formatted with the blocks as unlabelled columns and two rows for each genotype, named 
[line name]_1 and [line name]_2. The entries in this matrix are the calculated local 
GEBVs/block effects.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs.by.file}()},
\code{\link{save.pedigrees}()}
}
\concept{saving functions}
