% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.existing.groups}
\alias{see.existing.groups}
\title{Get a list of the groups currently existing in the SimData}
\usage{
see.existing.groups()
}
\value{
A dataframe containing two columns, the first, named "Group", 
being the group numbers of every group in the SimData that currently 
has members, the second, named "GroupSize", being 
the number of genotypes currently allocated to that group
}
\description{
Scans the saved data for groups that currently
have members and returns their group numbers and number of members.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()}
}
\concept{grouping functions}
