% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.genetic.map}
\alias{see.genetic.map}
\title{View one of the genetic maps being used by the simulation}
\usage{
see.genetic.map(mapID = 0L)
}
\arguments{
\item{mapID}{the identifier of the specific genetic map to view, or zero to 
view the first-loaded/default genetic map.}
}
\value{
A dataframe containing three columns: the first, named "marker", 
containing marker names; the second, named "chr", containing chromosome numbers;
and the third, named "pos", containing positions in cM.
}
\description{
Access the list of markers of one of the genetic maps loaded in 
the simulation. The row order of the returned dataframe will 
match the order of the markers as stored by the simulation (eg.
will match the marker order in genotypes from see.group.data(data.type="G")).
}
\details{
Note that genomicSimulation does not store chromosome names after initial
import. The "chr" column of the output contains the index of a chromosome
as stored in the simulation. 

Similarly, genomicSimulation does not store the marker positions exactly as 
they were in the input file. The first marker in a chromosome is designated 
to be at position 0, and the other markers are stored as their cM offset 
from the first marker's position. Chromosomes containing only one marker
do not track that marker's position, so that marker's position will be NaN 
in the output dataframe.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
