% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.group.data}
\alias{see.group.data}
\title{Get data of all individuals that belong to a group.}
\usage{
see.group.data(group, data.type)
}
\arguments{
\item{group}{an integer: the group number of the group to have a look at}

\item{data.type}{a string that will be used to identify the type of data
to be extracted. If the first character is 'N' or 'n',
the 'N'ames of group members will be extracted. If the first character is 
'D' or 'd', the I'D's of group members will be extracted. If the first character is
'I' or 'i', the 'I'ndexes of group members will be extracted. If the first character
is 'G' or 'g', the 'G'enotypes of the group members will be extracted.}
}
\value{
A vector containing the desired data output for each member of the group.
}
\description{
\code{see.group.data} allows you to extract the names or ids or indexes
or genotypes of group members as an R vector. Printing to a file then 
loading the data from the tab-separated text file is still the suggested
method for accessing data, but this may be used to peek at the data or 
make group adjustments during simulation time.
}
\details{
When loading genotypes as a group, some genotypes may have trailing nonsense
characters. Only the first (number of SNPs * 2) characters are valid. These trailing
characters occur because the C code does not allocate extra room for a null-terminating
byte, which makes the R function converting the genotypes to strings not know where the
end of the genotype is.
}
\section{Warning about indexes}{

Deleting a group (\code{\link{delete.group}}) may cause the indexes of group members to 
change (as the positions/indexes that they are stored in the internal matrix may be
rearranged). Therefore, you must ensure that there is no call to \code{\link{delete.group}}
between loading indexes with this function and passing them to a function that takes indexes
as inputs (eg \code{\link{make.group}}, \code{\link{cross}}).
}

\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.gebvs}()},
\code{\link{select.by.gebv}()}

Other data access functions: 
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.genotype}()}
}
\concept{data access functions}
\concept{grouping functions}
