% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.group.gebvs}
\alias{see.group.gebvs}
\title{Get GEBVs of a group}
\usage{
see.group.gebvs(group)
}
\arguments{
\item{group}{an integer: the group number of the group to return the GEBVs of.}
}
\value{
a dataframe whose columns are the GEBVs of the group members, named "GEBV", and 
the indexes of the group members, named "i".
}
\description{
\code{see.group.gebvs} gives the R user a dataframe containing the GEBVs and a matching
index for each genotype in a group. This can be used in conjunction with \code{\link{make.group}}
to perform a custom selection on a group.
}
\section{Usual workflow}{

The suggested workflow involves using this function to get GEBVs and matching indexes,
performing some transformation on the GEBVs (eg environmental masking then
sorting), identifying some of the entries for selection, and passing
the indexes of those entries to \code{\link{make.group}} in order to split
them off into a new group.
}

\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{select.by.gebv}()}

Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.genotype}()}
}
\concept{data access functions}
\concept{grouping functions}
