% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.group.gene.data}
\alias{see.group.gene.data}
\title{Get genotypes or allele counts of a group as a matrix.}
\usage{
see.group.gene.data(group, count.allele = NA_character_)
}
\arguments{
\item{group}{a vector of integers: the group number(s) of the group(s) to
access data from}

\item{count.allele}{a character that will be used to identify the type of data
to be extracted.
If it is NA, as it is by default, then return the pair of alleles as a string eg "AT"
If it is a character, then return the count of that allele eg. 'A' -> 0 if 
no copies of A at that marker, 1 if heterozygous for A, 2 if "AA" at that marker.}
}
\description{
\code{see.group.gene.data} allows you to extract a (markers)x(group members)
matrix of genotypes in simulation.
}
\details{
Multiple group numbers can be passed to this function. The function will still 
return a single matrix. Column representing all group members from the first
group will be first in the matrix, followed by columns representing all group
members in the second group, and so on. Groups are concatenated in the same order
as they were ordered in the input parameter passed to this function.
}
\seealso{
Other grouping functions: 
\code{\link{break.group.by.GEBV}()},
\code{\link{break.group.by.label.range}()},
\code{\link{break.group.by.label.value}()},
\code{\link{break.group.evenly}()},
\code{\link{break.group.into.buckets}()},
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.halfsib.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{break.group.randomly}()},
\code{\link{break.group.with.probabilities}()},
\code{\link{change.label.default}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()}

Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
\concept{grouping functions}
