% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.marker.effects}
\alias{see.marker.effects}
\title{View one of the sets of marker effects being used by the simulation}
\usage{
see.marker.effects(effect.set = 0L, effectID = 0L)
}
\arguments{
\item{effect.set}{the identifier of the specific marker effect set to view. 
If not specified, uses the first-loaded marker effect set in this simulation.
For legacy reasons, you can also set this parameter using the name `effectID`.}
}
\value{
A dataframe with columns "marker", "allele" and "eff"
}
\description{
The function 
will return a dataframe with three columns: 
the first, named "marker", containing marker names; 
the second, named "allele", containing an allele symbol;
and the third, named "eff", containing the contribution of that allele at
that marker to the estimated breeding value.
}
\details{
If the set of marker effects has centering values, then rows containing the
special value "(centre)" in the allele column will contain that marker's 
centring value.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
