% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.minimal.GEBV}
\alias{see.minimal.GEBV}
\alias{see.minimal.gebv}
\title{Get the lowest-possible GEBV given the current loaded effect values.}
\usage{
see.minimal.GEBV(eff.set = 1L)
}
\arguments{
\item{eff.set}{identifier for the set of marker effects to use to calculate the
minimum GEBV.}
}
\value{
The lowest possible GEBV
}
\description{
\code{see.minimal.GEBV} allows you to extract the lowest GEBV score
possible according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
