% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data-access.R
\name{see.minimum.GEBV}
\alias{see.minimum.GEBV}
\title{Get the lowest-possible GEBV given the current loaded effect values.}
\usage{
see.minimum.GEBV()
}
\value{
The lowest possible GEBV
}
\description{
\code{see.minimum.GEBV} allows you to extract the lowest GEBV score
possible according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()}
}
\concept{data access functions}
