% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{see.optimal.GEBV}
\alias{see.optimal.GEBV}
\title{Get the ultimate/highest-possible GEBV given the current loaded values.}
\usage{
see.optimal.GEBV()
}
\value{
The highest possible GEBV
}
\description{
\code{see.optimal.GEBV} allows you to extract the optimal GEBV
according to the current loaded effect values as an R string. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other saving functions: 
\code{\link{save.GEBVs}()},
\code{\link{save.allele.counts}()},
\code{\link{save.genome.model}()},
\code{\link{save.genotypes}()},
\code{\link{save.local.GEBVs}()},
\code{\link{save.pedigrees}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()},
\code{\link{see.optimal.genotype}()}
}
\concept{saving functions}
