% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.optimal.GEBV}
\alias{see.optimal.GEBV}
\alias{see.optimal.gebv}
\title{Get the ultimate/highest-possible GEBV given the current loaded effect values.}
\usage{
see.optimal.GEBV(eff.set = 1L)
}
\arguments{
\item{eff.set}{identifier for the set of marker effects to use to calculate the
maximal GEBV}
}
\value{
The highest possible GEBV
}
\description{
Allows you to extract the optimal GEBV
according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
