% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.optimal.GEBV}
\alias{see.optimal.GEBV}
\alias{see.optimal.gebv}
\title{Get the ultimate/highest-possible GEBV given the current loaded effect values.}
\usage{
see.optimal.GEBV(effect.set = 0L, eff.set = 0L)
}
\arguments{
\item{effect.set}{identifier for the set of marker effects to use to calculate the
maximal GEBV. If not otherwise specified, uses the earliest-loaded marker 
effect set in this simulation.
For legacy reasons, you can also set this parameter using the name `eff.set`.}

\item{eff.set}{Legacy name for the parameter now called `effect.set`}
}
\value{
The highest possible GEBV
}
\description{
Allows you to extract the optimal GEBV
according to the current loaded effect values. An error is 
raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.GEBVs}()},
\code{\link{see.genetic.map}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.local.GEBVs}()},
\code{\link{see.marker.effects}()},
\code{\link{see.markerblocks}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()},
\code{\link{see.optimal.possible.haplotype}()}
}
\concept{data access functions}
