% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-calculators.R
\name{see.optimal.possible.haplotype}
\alias{see.optimal.possible.haplotype}
\title{Get a string containing the ultimate/highest-scoring set of alleles available
in the current group.}
\usage{
see.optimal.possible.haplotype(group, eff.set = 1L)
}
\arguments{
\item{group}{an integer: the group number of the group to have a look at. 
Can be a vector of groups, in which case the calculation will be run independently
for each group.}

\item{eff.set}{identifier for the set of marker effects to use to calculate the
optimal haplotype and haplotype breeding values}
}
\value{
A string containing the allele out of all alleles in the group
that has the highest effect value at that locus. The string will be ordered in
genome order (lowest chromosome and lowest position to highest) according
to the map that was included on initialisation.
}
\description{
That is, consider the pool of alleles that exist in the group for each locus,
and take the highest-scoring allele at each locus.
}
\details{
An error is raised if no effect values have been loaded.
}
\seealso{
Other data access functions: 
\code{\link{see.group.data}()},
\code{\link{see.group.gene.data}()},
\code{\link{see.minimal.GEBV}()},
\code{\link{see.optimal.GEBV}()},
\code{\link{see.optimal.haplotype}()},
\code{\link{see.optimal.possible.GEBV}()}
}
\concept{data access functions}
