% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-group-utils.R
\name{select.by.gebv}
\alias{select.by.gebv}
\title{Perform selection on a group by true GEBV.}
\usage{
select.by.gebv(
  from.group,
  low.score.best = FALSE,
  percentage = NULL,
  number = NULL
)
}
\arguments{
\item{from.group}{an integer: the group number of the group perform selection on.
If a vector, the selection will be done on each group in the vector.}

\item{low.score.best}{If FALSE, then the highest GEBVs are selected. If TRUE,
then the lower ones are selected.}

\item{percentage}{If this is a number and number and threshold are NULL, then 
the [percentage]\% with the best GEBVs will be selected into the new group. Note that
this is a whole-number percentage, i.e. for 5\%, set percentage=5 not 0.05.}

\item{number}{If this is an integer and percentage and threshold are NULL, then
the [number] with the best GEBVs will be selected into the new group.}
}
\value{
the group number of the new group, containing the genotypes that were 
positively selected.
}
\description{
\code{select.by.gebv} calculates the GEBVs of all members of a group, sorts them
in either ascending or descending order, and then selects the best few. The number
selected can be an exact number (select the best 5) or a percentage (take the best
2\%). Exactly one of percentage and number parameters must be defined for the
function to run.
}
\section{Custom selection}{

For more complex selection than by true/unmasked GEBV, use the function
\code{\link{see.group.gebvs}}, use R scripting to identify which to select,
and split them into a new group using \code{\link{make.group}}.
}

\seealso{
Other grouping functions: 
\code{\link{break.group.into.families}()},
\code{\link{break.group.into.individuals}()},
\code{\link{combine.groups}()},
\code{\link{delete.group}()},
\code{\link{make.group}()},
\code{\link{see.existing.groups}()},
\code{\link{see.group.data}()},
\code{\link{see.group.gebvs}()}
}
\concept{grouping functions}
